requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
filelock==3.10.7
    # via torch
hydra-core==1.3.2
    # via SpoofingDetection (setup.py)
idna==3.4
    # via requests
jinja2==3.1.2
    # via torch
markupsafe==2.1.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.1
    # via torch
numpy==1.24.2
    # via
    #   opencv-python
    #   torchvision
omegaconf==2.3.0
    # via
    #   SpoofingDetection (setup.py)
    #   hydra-core
opencv-python==4.7.0.72
    # via SpoofingDetection (setup.py)
packaging==23.0
    # via hydra-core
pillow==9.5.0
    # via torchvision
pyyaml==6.0
    # via omegaconf
requests==2.28.2
    # via torchvision
sympy==1.11.1
    # via torch
torch==2.0.0
    # via
    #   SpoofingDetection (setup.py)
    #   torchvision
torchvision==0.15.1
    # via SpoofingDetection (setup.py)
typing-extensions==4.5.0
    # via torch
urllib3==1.26.15
    # via requests
